move buildGraph to utility
